+++ APPartment.API/Meta/Controllers/CommentsController.cs
@@ -41,7 +41,7 @@ namespace APPartment.API.Controllers
             try
             {
                 comment = BaseCRUDService.Save(comment);
-                BaseCRUDService.AddUserAsParticipantToObject(comment.TargetObjectID, CurrentUserID, comment.ObjectTypeID);
+                BaseCRUDService.AddUserAsParticipantToObject(comment.TargetObjectID, (long)CurrentUserID, comment.ObjectTypeID);
                 var result = new CommentsRenderer(CurrentUserID, CurrentHomeID).BuildPostComment(comment);
 
                 return Ok(result);
